edx_dl: Print video IDs as soon as we discover them.
This way we don't have to accumulate all the video IDs before printing them
all in one go.  This is specially nice if, for some reason, the script dies
(say, due to many connections to the remote site).

One speedup that was implemented in coursera-dl that is applicable here is
the migration to the python library called requests, which actually keeps
state between many requests and is likely to save bandwidth and also be good
netizens.

Caching the cookies is also a great feature that was implemented there and
that should also speedup the parsing of the pages.